Move longlink validator and input validator into their own file
