+++ app/usecase/input/longlink_test.go
@@ -1,10 +1,9 @@
 package input
 
 import (
+	"github.com/stretchr/testify/assert"
 	"strings"
 	"testing"
-
-	"github.com/stretchr/testify/assert"
 )
 
 func TestLongLink_IsValid(t *testing.T) {
@@ -46,42 +45,10 @@ func TestLongLink_IsValid(t *testing.T) {
 		},
 	}
 
-	validator := NewLongLinkValidator()
+	validator := NewLongLink()
 	for _, testCase := range testCases {
 		t.Run(testCase.name, func(t *testing.T) {
 			assert.Equal(t, testCase.expIsValid, validator.IsValid(&testCase.longLink))
-		})
-	}
-}
-
-func TestCustomAlias_IsValid(t *testing.T) {
-	t.Parallel()
-	testCases := []struct {
-		name       string
-		alias      string
-		expIsValid bool
-	}{
-		{
-			name:       "empty string",
-			alias:      "",
-			expIsValid: true,
-		},
-		{
-			name:       "alias too long",
-			alias:      strings.Repeat("helloworld", 5),
-			expIsValid: false,
-		},
-		{
-			name:       "alias valid",
-			alias:      "fb",
-			expIsValid: true,
-		},
-	}
-
-	validator := NewCustomAliasValidator()
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			assert.Equal(t, testCase.expIsValid, validator.IsValid(&testCase.alias))
 		})
 	}
 }
